--- app/components/dashboard/Usage.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-const Usage = () => {
-  return <div>Usage</div>;
-};
-
-export default Usage;
